Add files via upload
In this project, it is requested to develop a transportation database model in which different transportation alternatives for a country (sea, air, road, mobile transportation, metro, etc.) are recorded at macro and micro levels. In this system, transportation will be provided with a unique card for each citizen. Citizens' transportation histories, searches and preferences will be recorded. Macro and micro level queries can be made in the system.